move newcourse page to folder
